+++ ReleaseNotes/release_notes.md
@@ -2,9 +2,12 @@
 
    1. slightly improve printts from for `Files/Data/...` - make them a bit more informative
    1. add `truncate_middle` function into `ostap.utils.strings` module: A little bit modified version of the code from the `pidgen2` project by Anton Pluektov
-    
+   1. add `ResoNeedham` as explicit resolution function
+     
 ## Backward-incompatible 
 
+   1.  Rename `a0`, `a1` and `a2` parameterd of Needham's function to `c0`, `c1` and `c2`. Update `Ostap::Math::Needham`, `Ostap::Models::Needham` and `Needham_pdf` `
+    
 ## Bug fixes
 
 # v3.0.1.2